Updated Property Profit Calculator to Take Input from Webpage
Now the monthly profit calculator uses streamlit to take necessary user inputs from a webpage and output a downloadable spreadsheet instead of writing it to a local file. This is done on localhost right now, no live website has been created for this yet.

Added more accurate TODO comments to reflect my intentions for the future of this project.

Added some basic debugging code. This will be removed or commented out later.