Adding cash flow calculation to property profit calculator
The property profit calculator now checks month by month to see if a rental property is cash flowing or not for its new owner on a monthly basis.

If it does, it will state that it does on the webpage and let the owner know which month it begins to cash flow and by how much.

If it does not, it lets the owner know that it does not and how much the owner is losing per month in the last month.