Fixed Visual Issues with GIS/CPP Optimizer
Switched the display order of the 2 generated graphs.

Rounded the numbers in the dataframe used for the CSV to 2 decimal places.

Fixed an issue where pressing the button to generate a CSV file representing the data would hide the generated optimized data in the webpage.

Cleaned up some of the code.