Large changes to Property Profit Calculator
Refactoring was done such that the backend file is 1 file instead of 4. The column renaming was also moved to the backend. References to "house" were renamed "property".

Inputs were changed from text_input to number_input for cleaner error and help messages for the user.

Plots and table updated such that the table and plots only show relevant data based on the user inputs. There is a bug here that needs to be addressed where sometimes the table adds extra columns.

A small change was also made to the GIS/CPP optimizer removing many of the step requirements for user inputs.