Added Variable Mortgage Loan Terms to Property Profit Calculator
Added functionality to pull in variable loan terms with different interest rates to simulate changing interest rates over time.

Users can only do this on a year by year basis with this change, but can have as money loan terms as they want so long as the sum of their years is equal to the amortization period.

This could be updated to go month by month later.

The backend calculations were updated in kind, and the spreadsheet has a new column to keep track of the changing interest rate over time.